mv getRoutePathParams() helper.
