+++ express-zod-api/tests/common-helpers.spec.ts
@@ -10,6 +10,7 @@ import {
   makeCleanId,
   ensureError,
   pullExampleProps,
+  getRoutePathParams,
 } from "../src/common-helpers";
 import { z } from "zod";
 import { makeRequestMock } from "../src/testing";
@@ -18,6 +19,34 @@ describe("Common Helpers", () => {
   describe("defaultInputSources", () => {
     test("should be declared in a certain way", () => {
       expect(defaultInputSources).toMatchSnapshot();
+    });
+  });
+
+  describe("getRoutePathParams()", () => {
+    test("should return an array of param names", () => {
+      expect(getRoutePathParams("/users/:userId/books/:bookId")).toEqual([
+        "userId",
+        "bookId",
+      ]);
+      expect(getRoutePathParams("/flights/:from-:to")).toEqual(["from", "to"]);
+      expect(getRoutePathParams("/something")).toEqual([]);
+      expect(getRoutePathParams("")).toEqual([]);
+      expect(getRoutePathParams("\n")).toEqual([]);
+    });
+
+    test("should return an array of param names", () => {
+      expect(getRoutePathParams("/users/:userId/books/:bookId")).toEqual([
+        "userId",
+        "bookId",
+      ]);
+      expect(getRoutePathParams("/flights/:from-:to")).toEqual(["from", "to"]);
+      expect(getRoutePathParams("/test/:genus.:species")).toEqual([
+        "genus",
+        "species",
+      ]);
+      expect(getRoutePathParams("/something")).toEqual([]);
+      expect(getRoutePathParams("")).toEqual([]);
+      expect(getRoutePathParams("\n")).toEqual([]);
     });
   });
